Math.round() needs float variable [Findbugs]
git-svn-id: https://svn.apache.org/repos/asf/jakarta/jmeter/branches/rel-2-1@408507 13f79535-47bb-0310-9956-ffa450edef68

Former-commit-id: 3f69d9cddd20546c6c3e7e24d91a8be94091bfa2

+++ src/core/org/apache/jmeter/reporters/FileReporter.java
@@ -323,7 +323,7 @@ private static class GraphPanel extends JPanel {
 		// draw grid
 		g.setColor(Color.gray);
 		int dataWidth = getDataWidth();
-		int increment = Math.round((width - 1) / (dataWidth - 1));
+		int increment = Math.round((float)(width - 1) / (dataWidth - 1));
 
 		/*
 		 * for (int t = 0; t < dataWidth; t++) { g.drawLine(t * increment, 0, t *
